+++ examples/discovery-search-app/CHANGELOG.md
@@ -10,6 +10,13 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+## [1.0.7-rc.1](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.7-rc.0...discovery-search-app@1.0.7-rc.1) (2020-01-15)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
 
 ## [1.0.2-beta.20](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.2-beta.19...discovery-search-app@1.0.2-beta.20) (2020-01-16)
 
@@ -175,6 +182,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 ## 1.0.2-beta.0 (2019-12-09)
+## 1.0.7-rc.0 (2020-01-14)
 
 **Note:** Version bump only for package discovery-search-app