+++ packages/discovery-react-components/CHANGELOG.md
@@ -109,6 +109,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ### Bug Fixes
 
 * error in document preview with split docs ([#29](https://github.com/watson-developer-cloud/discovery-components/issues/29)) ([ed47956](https://github.com/watson-developer-cloud/discovery-components/commit/ed47956))
+## [1.0.7-rc.1](https://github.com/watson-developer-cloud/discovery-components/compare/@ibm-watson/discovery-react-components@1.0.7-rc.0...@ibm-watson/discovery-react-components@1.0.7-rc.1) (2020-01-15)
 
 
 ### Features
@@ -116,12 +117,14 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 * structured query add and remove rules ([#27](https://github.com/watson-developer-cloud/discovery-components/issues/27)) ([eab1d3b](https://github.com/watson-developer-cloud/discovery-components/commit/eab1d3b))
 * structured query basic ([#22](https://github.com/watson-developer-cloud/discovery-components/issues/22)) ([b992f63](https://github.com/watson-developer-cloud/discovery-components/commit/b992f63))
 * upgrade storybook to 5.3 ([#25](https://github.com/watson-developer-cloud/discovery-components/issues/25)) ([93db599](https://github.com/watson-developer-cloud/discovery-components/commit/93db599))
+* allow custom document text field name ([#55](https://github.com/watson-developer-cloud/discovery-components/issues/55)) ([c24363f](https://github.com/watson-developer-cloud/discovery-components/commit/c24363f))
 
 
 
 
 
 # [1.1.0-beta.3](https://github.com/watson-developer-cloud/discovery-components/compare/@ibm-watson/discovery-react-components@1.0.7-beta.5...@ibm-watson/discovery-react-components@1.1.0-beta.3) (2019-12-23)
+## [1.0.7-rc.0](https://github.com/watson-developer-cloud/discovery-components/compare/@ibm-watson/discovery-react-components@1.0.6...@ibm-watson/discovery-react-components@1.0.7-rc.0) (2020-01-14)
 
 
 ### Bug Fixes
@@ -140,6 +143,9 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 # [1.1.0-beta.2](https://github.com/watson-developer-cloud/discovery-components/compare/@ibm-watson/discovery-react-components@1.0.7-beta.5...@ibm-watson/discovery-react-components@1.1.0-beta.2) (2019-12-20)
+* issue with document highlighting offset ([#46](https://github.com/watson-developer-cloud/discovery-components/issues/46)) ([9eab866](https://github.com/watson-developer-cloud/discovery-components/commit/9eab866))
+* make DocPrev only display passage text with highlighting ([#50](https://github.com/watson-developer-cloud/discovery-components/issues/50)) ([51fb55a](https://github.com/watson-developer-cloud/discovery-components/commit/51fb55a))
+* pr to add PR13 into janFixpack branch ([#47](https://github.com/watson-developer-cloud/discovery-components/issues/47)) ([bbdb883](https://github.com/watson-developer-cloud/discovery-components/commit/bbdb883))
 
 
 ### Features
@@ -226,6 +232,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ### Bug Fixes
 
 * **SearchResults:** title fallback and footerbar value max length ([#9](https://github.com/watson-developer-cloud/discovery-components/issues/9)) ([7ea3ff1](https://github.com/watson-developer-cloud/discovery-components/commit/7ea3ff1))
+* standalone HTML view with table highlighting ([#48](https://github.com/watson-developer-cloud/discovery-components/issues/48)) ([d8b1207](https://github.com/watson-developer-cloud/discovery-components/commit/d8b1207))