+++ packages/discovery-styles/CHANGELOG.md
@@ -81,6 +81,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 ## [1.0.7-beta.3](https://github.com/watson-developer-cloud/discovery-components/compare/@ibm-watson/discovery-styles@1.0.6...@ibm-watson/discovery-styles@1.0.7-beta.3) (2019-12-11)
+## [1.0.7-rc.1](https://github.com/watson-developer-cloud/discovery-components/compare/@ibm-watson/discovery-styles@1.0.7-rc.0...@ibm-watson/discovery-styles@1.0.7-rc.1) (2020-01-15)
 
 **Note:** Version bump only for package @ibm-watson/discovery-styles
 
@@ -110,6 +111,12 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ### Bug Fixes
 
 * **SearchResults:** title fallback and footerbar value max length ([#9](https://github.com/watson-developer-cloud/discovery-components/issues/9)) ([7ea3ff1](https://github.com/watson-developer-cloud/discovery-components/commit/7ea3ff1))
+## [1.0.7-rc.0](https://github.com/watson-developer-cloud/discovery-components/compare/@ibm-watson/discovery-styles@1.0.6...@ibm-watson/discovery-styles@1.0.7-rc.0) (2020-01-14)
+
+
+### Features
+
+* standalone HTML view with table highlighting ([#48](https://github.com/watson-developer-cloud/discovery-components/issues/48)) ([d8b1207](https://github.com/watson-developer-cloud/discovery-components/commit/d8b1207))